Missing crossorigin in html output
I can see the support of crossorigin attribute in header but not in the html ouput :    
/vendor/yireo/magento2-linkpreload/Link/Link.php
function getHeader

But in the navigator, I have a notice in web developer :
A preload for '../Blank-Theme-Icons/Blank-Theme-Icons.woff2' is found, but is not used because the request credentials mode does not match. Consider taking a look at crossorigin attribute.

This is my proposal to fix this issue.